Moving CSV translator into Translators module
