+++ lib/rof.rb
@@ -5,7 +5,7 @@ require "rof/access"
 require "rof/collection"
 require "rof/utility"
 require "rof/rdf_context"
-require "rof/translate_csv"
+require "rof/translators/translate_csv"
 require "rof/filters/date_stamp"
 require "rof/filters/file_to_url"
 require "rof/filters/label"
