+++ spec/lib/rof/translators/translate_csv_spec.rb
@@ -1,13 +1,13 @@
 require 'spec_helper'
 
-module ROF
+module ROF::Translators
   describe "translate CSV" do
     it "requires the columns type and owner" do
       s = "dc:title,access,owner"
-      expect{TranslateCSV.run(s)}.to raise_error(ROF::TranslateCSV::MissingOwnerOrType)
+      expect{TranslateCSV.run(s)}.to raise_error(ROF::Translators::TranslateCSV::MissingOwnerOrType)
 
       s = "dc:title,access,type"
-      expect{TranslateCSV.run(s)}.to raise_error(ROF::TranslateCSV::MissingOwnerOrType)
+      expect{TranslateCSV.run(s)}.to raise_error(ROF::Translators::TranslateCSV::MissingOwnerOrType)
 
       s = "dc:title,type,owner,access"
       expect(TranslateCSV.run(s)).to eq([])
@@ -17,7 +17,7 @@ module ROF
       s = %q{type,owner
       Work,
       }
-      expect{TranslateCSV.run(s)}.to raise_error(ROF::TranslateCSV::MissingOwnerOrType)
+      expect{TranslateCSV.run(s)}.to raise_error(ROF::Translators::TranslateCSV::MissingOwnerOrType)
     end
 
     it "deocdes the access field into rights" do
@@ -38,7 +38,7 @@ module ROF
         "owner" => "user1",
         "rights" => {"edit" => ["user1"]},
         "metadata" => {
-          "@context" => RdfContext,
+          "@context" => ROF::RdfContext,
           "dc:title" => "Q, A Letter",
           "foaf:name" => ["Jane Smith", "Zander"]}
       }])
@@ -67,7 +67,7 @@ module ROF
         "owner" => "user1",
         "rights" => {"edit" => ["user1"]},
         "metadata" => {
-          "@context" => RdfContext,
+          "@context" => ROF::RdfContext,
           "dc:title" => "Q, A Letter",
           "foaf:name" => ["Jane Smith", "Zander"]}
       }])
@@ -84,7 +84,7 @@ module ROF
         "owner" => "user1",
         "rights" => {"edit" => ["user1"]},
         "metadata" => {
-          "@context" => RdfContext,
+          "@context" => ROF::RdfContext,
           "dc:title" => "Q, A Letter"},
         "files" => [
           "thumb",
@@ -101,7 +101,7 @@ module ROF
       s = %q{type,owner,dc:title,files
       +,user1,,extra file.txt
       }
-      expect {TranslateCSV.run(s)}.to raise_error(ROF::TranslateCSV::NoPriorWork)
+      expect {TranslateCSV.run(s)}.to raise_error(ROF::Translators::TranslateCSV::NoPriorWork)
     end
